+++ apps/api/src/graphql/job-application/resolvers/job-application/inputs/update-job-application-input.ts
@@ -1,0 +1,17 @@
+import {Field, InputType} from "type-graphql";
+import {JobApplicationStatus} from "@graphql/job-application/enums/job-application-status";
+
+@InputType()
+export class UpdateJobApplicationInput {
+  @Field()
+  id: string;
+
+  @Field({ nullable: true})
+  url?: string;
+
+  @Field({ nullable: true})
+  title?: string;
+
+  @Field(() => JobApplicationStatus, { nullable: true})
+  status?: JobApplicationStatus;
+}