+++ apps/api/src/graphql/job-application/resolvers/job-application/mutation-resolver.ts
@@ -1,10 +1,13 @@
 import {Arg, Ctx, Mutation, Resolver} from "type-graphql";
 import {JobApplication} from "@graphql/job-application/entities/job-application";
-import type  {myContext} from "@/utils/create-server";
+import type {myContext} from "@/utils/create-server";
 import {
   CreateJobApplicationInput
 } from "@graphql/job-application/resolvers/job-application/inputs/create-job-application-input";
 import {GraphQLError} from "graphql/error";
+import {
+  UpdateJobApplicationInput
+} from "@graphql/job-application/resolvers/job-application/inputs/update-job-application-input";
 
 @Resolver(() => JobApplication)
 export class JobApplicationMutationResolver {
@@ -19,8 +22,6 @@ export class JobApplicationMutationResolver {
 
     if (findJobApplication) throw new GraphQLError('Job application url already exists');
 
-
-
     const jobApplication = em.create(JobApplication, {
       url: input.url,
       status: input.status,
@@ -29,5 +30,36 @@ export class JobApplicationMutationResolver {
     await em.persistAndFlush(jobApplication);
 
     return jobApplication;
+  }
+
+  @Mutation(() => JobApplication)
+  async updateJobApplication(
+    @Arg('input') input: UpdateJobApplicationInput,
+    @Ctx() {em}: myContext
+  ) {
+    let jobApplication = await em.findOne(JobApplication, {id: input.id});
+    if (!jobApplication) throw new GraphQLError('Job application not found');
+
+   if (input.url) jobApplication.url = input.url;
+    if (input.status) jobApplication.status = input.status;
+    if (input.title) jobApplication.title = input.title;
+
+    await em.persistAndFlush(jobApplication);
+
+    return jobApplication;
+  }
+
+  @Mutation(() => Boolean)
+  async deleteJobApplication(
+    @Arg('id') id: string,
+    @Ctx() {em}: myContext
+  ) {
+    let jobApplication = await em.findOne(JobApplication, {id});
+    if (!jobApplication) throw new GraphQLError('Job application not found');
+
+    jobApplication.deletedAt = new Date();
+    await em.persistAndFlush(jobApplication);
+
+    return true;
   }
 }