+++ packages/codegen-api/src/graphql.ts
@@ -13,6 +13,8 @@ export type Scalars = {
   Boolean: { input: boolean; output: boolean; }
   Int: { input: number; output: number; }
   Float: { input: number; output: number; }
+  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar.This scalar is serialized to a string in ISO 8601 format and parsed from a string in ISO 8601 format. */
+  DateTimeISO: { input: any; output: any; }
 };
 
 export type CreateJobApplicationInput = {
@@ -23,6 +25,7 @@ export type CreateJobApplicationInput = {
 
 export type JobApplication = {
   __typename?: 'JobApplication';
+  deletedAt?: Maybe<Scalars['DateTimeISO']['output']>;
   id: Scalars['String']['output'];
   status: JobApplicationStatus;
   title: Scalars['String']['output'];
@@ -39,6 +42,8 @@ export enum JobApplicationStatus {
 export type Mutation = {
   __typename?: 'Mutation';
   createJobApplication: JobApplication;
+  deleteJobApplication: Scalars['Boolean']['output'];
+  updateJobApplication: JobApplication;
 };
 
 
@@ -46,7 +51,24 @@ export type MutationCreateJobApplicationArgs = {
   input: CreateJobApplicationInput;
 };
 
+
+export type MutationDeleteJobApplicationArgs = {
+  id: Scalars['String']['input'];
+};
+
+
+export type MutationUpdateJobApplicationArgs = {
+  input: UpdateJobApplicationInput;
+};
+
 export type Query = {
   __typename?: 'Query';
   jobApplications: Array<JobApplication>;
+};
+
+export type UpdateJobApplicationInput = {
+  id: Scalars['String']['input'];
+  status?: InputMaybe<JobApplicationStatus>;
+  title?: InputMaybe<Scalars['String']['input']>;
+  url?: InputMaybe<Scalars['String']['input']>;
 };