+++ apps/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web
 
+## 0.2.0
+
+### Minor Changes
+
+- added delete button
+  enabled update status button
+
 ## 0.1.1
 
 ### Patch Changes